Forces bit size of IEC_TIMESPEC. "long int" size isn't the same on 32b and 64b targets.

+++ lib/C/iec_types.h
@@ -35,8 +35,8 @@ typedef double   IEC_LREAL;
  *          (these macros may be found in iec_std_lib.h)
  */
 typedef struct {
-    long int tv_sec;            /* Seconds.  */
-    long int tv_nsec;           /* Nanoseconds.  */
+    int64_t tv_sec;            /* Seconds. y2038 ready */
+    int32_t tv_nsec;           /* Nanoseconds. 32b is enough */
 } /* __attribute__((packed)) */ IEC_TIMESPEC;  /* packed is gcc specific! */
 
 typedef IEC_TIMESPEC IEC_TIME;
